Some shell script improvements

+++ import/import.sh
@@ -6,17 +6,33 @@
 # See http://wiki.openstreetmap.org/wiki/OpenLinkMap for details.
 
 
-# working directory, please change
-cd /home/www/sites/194.245.35.149/site/olm/import
-PATH="$PATH:/home/www/sites/194.245.35.149/site/olm/import/bin"
+# path to the project directory
+PROJECTPATH=/home/www/sites/194.245.35.149/site/olm
+# directory where the planet file and other data files are stored, can be equal to PROJECTPATH
+DATAPATH=/home/www/sites/194.245.35.149/site/olm/import
+PATH="$PATH:$DATAPATH/bin"
+PATH="$PATH:$PROJECTPATH/import/bin/osm2pgsql"
 export JAVACMD_OPTIONS=-Xmx2800M
 
+cd $DATAPATH
 
-# download planet file, ~ 8 hours
-echo "Downloading planet file"
+
+# download planet file if not existing
+echo "Getting planet file if necessary"
 echo ""
-wget http://planet.openstreetmap.org/pbf/planet-latest.osm.pbf
-mv planet-latest.osm.pbf old.pbf
+if [ ! -f old.pbf ]; then
+	echo "Planet file not existing, now downloading it"
+	wget http://planet.openstreetmap.org/pbf/planet-latest.osm.pbf
+	mv planet-latest.osm.pbf old.pbf
+	# update planet file
+	echo "Updating planet file"
+	echo ""
+	date -u +%s > timestamp
+	osmupdate old.pbf new.pbf --max-merge=2 --hourly --drop-author -v
+	rm old.pbf
+	mv new.pbf old.pbf
+	echo ""
+fi
 echo ""
 
 
@@ -30,14 +46,14 @@ mv new.pbf old.pbf
 echo ""
 
 
-# convert planet file, ~ 25 min
+# convert planet file
 echo "Converting planet file"
 echo ""
 osmconvert old.pbf --drop-relations --out-o5m >temp.o5m
 echo ""
 
 
-# filter planet file, ~ 35 min
+# filter planet file
 echo "Filtering planet file"
 echo ""
 osmfilter temp.o5m --keep="wikipedia= wikipedia:*= contact:phone= website= url= phone= fax= email= addr:email= image= url:official= contact:website= addr:phone= phone:mobile= contact:mobile= addr:fax= contact:email= contact:fax= image:panorama= opening_hours=" --out-o5m >temp-olm.o5m
@@ -47,7 +63,7 @@ rm temp.o5m
 echo ""
 
 
-# create centroids, remove not-node elements, ~ 2 min
+# create centroids, remove not-node elements
 echo "Creating centroids, removing elements"
 echo ""
 osmconvert temp-olm.o5m --all-to-nodes --max-objects=90000000 --out-o5m >temp.o5m
@@ -74,7 +90,7 @@ rm old-nextobjects.pbf
 echo ""
 
 
-# importing in database, ~ 20 min
+# importing in database
 echo "Importing in database"
 echo ""
 php import.php
